requirements.txt
flask
qrcode
google-generativeai
firebase_admin
Pillow